cleanup generator
