+++ packages/widgetbook_generator/lib/extensions/json_list_formatter.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-import 'models/widgetbook_data.dart';
+import '../models/widgetbook_data.dart';
 
-extension JsonListExtension on List<WidgetbookData> {
+extension JsonListFormatter on List<WidgetbookData> {
   String toJson() => const JsonEncoder.withIndent('  ')
       .convert(map((data) => data.toJson()).toList());
 }
