+++ packages/widgetbook_generator/lib/generators/json_builder.dart
