+++ packages/widgetbook_generator/lib/generators/tree_service.dart
