+++ packages/widgetbook_generator/lib/instances/base_instance.dart
