+++ packages/widgetbook_generator/lib/instances/boolean_instance.dart
