+++ packages/widgetbook_generator/lib/instances/lambda_instance.dart
