+++ packages/widgetbook_generator/lib/instances/string_instance.dart
