+++ packages/widgetbook_generator/test/src/instances/primary_instance_test.dart
@@ -1,6 +1,6 @@
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
-import 'package:widgetbook_generator/code_generators/instances/primary_instance.dart';
+import 'package:widgetbook_generator/instances/primary_instance.dart';
 
 class CustomType {}
