+++ packages/widgetbook_generator/test/src/instances/string_instance_test.dart
@@ -1,6 +1,6 @@
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
-import 'package:widgetbook_generator/code_generators/instances/string_instance.dart';
+import 'package:widgetbook_generator/instances/string_instance.dart';
 
 void main() {
   group(
