+++ packages/widgetbook_generator/test/src/instances/widgetbook_widget_instance_test.dart
@@ -1,11 +1,11 @@
 import 'package:test/test.dart';
-import 'package:widgetbook_generator/code_generators/instances/list_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/widgetbook_use_case_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/widgetbook_widget_instance.dart';
-import 'package:widgetbook_generator/code_generators/properties/property.dart';
+import 'package:widgetbook_generator/instances/list_instance.dart';
+import 'package:widgetbook_generator/instances/property.dart';
+import 'package:widgetbook_generator/instances/widgetbook_use_case_instance.dart';
+import 'package:widgetbook_generator/instances/widgetbook_widget_instance.dart';
 import 'package:widgetbook_generator/models/widgetbook_use_case_data.dart';
 
-import '../instance_helper.dart';
+import 'instance_helper.dart';
 
 void main() {
   group(
